Invert brief ids and supplier ids
Up to this point, we have found the briefs with updated questions
and answers and then we have looked for suppliers who are interested
in them.

When we send the emails, we need to send one to each supplier
with links to the briefs that have been updated.  This means that
we need to invert our collection from:

- brief ids pointing to a list of interested supplier ids
to
- supplier ids pointing to a list of briefs are interested in