Add branch to clean up agreements in bad state
If a supplier created a new framework agreement in the time between the
generate script being run and the upload script being run, we will have
uploaded a possible outdated framework agreement PDF and attached it to
the wrong framework agreement object. This means the agreement object
will have status "countersigned" but not have a "countersignedAt" entry.

We know of at least one supplier who this happened to, we should clean
up others until we close this race condition.