Calculate top words in support tickets
This jupyter notebook calculates top words for a dataset of support requests, using tf-idf.

Conclusion:
---------
Since support tickets are fairly short documents, and we have a small dataset (100s of requests), using tf-idf out of the box rarely gives good guesses.
In this case, manually classifying and defining a taxonomy is a better strategy.

To try to classify programmatically we would need:
- a larger set of training data
- custom stop words set to eliminate terms like https or uk
- investigate classification and clustering algorithms
- potentially consider multi-term expressions like "digital marketplace"